requirements.txt
opencv-python==4.8.1.78
Pillow==10.0.1
Flask==2.3.3
Flask-CORS==4.0.0
numpy==1.24.3
textstat==0.7.3
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.3